requirements.txt
sentence-transformers
numpy
pandas
scikit-learn
faiss-cpu
fastapi
uvicorn
python-multipart
streamlit
nltk